+++ nemesis/src/response.rs
@@ -9,35 +9,21 @@ use log::warn;
 use mruby::interpreter::Mrb;
 use mruby::value::{Value, ValueLike};
 use mruby::MrbError;
-use rocket::http::Status;
-use rocket::request::Request;
-use rocket::response::{self, Responder};
 use std::collections::HashMap;
 use std::convert::{self, TryFrom};
-use std::io::Cursor;
 
 use crate::nemesis;
 use crate::Error;
 
 #[derive(Debug)]
 pub struct Response {
-    status: Status,
-    headers: HashMap<String, String>,
-    body: Vec<u8>,
+    pub status: u16,
+    pub headers: HashMap<String, String>,
+    pub body: Vec<u8>,
 }
 
 impl Response {
     const RACK_RESPONSE_TUPLE_LEN: usize = 3;
-
-    pub fn into_rocket<'a>(self) -> rocket::Response<'a> {
-        let mut response = rocket::Response::build();
-        response.status(self.status);
-        response.sized_body(Cursor::new(self.body));
-        for (key, value) in self.headers {
-            response.raw_header(key, value);
-        }
-        response.finalize()
-    }
 
     /// Convert from a Rack `[status, headers, body]` response tuple to a Rust
     /// representation. This code converts a response tuple using the Ruby class
@@ -63,10 +49,9 @@ impl Response {
         })
     }
 
-    fn status(response: &Value) -> Result<Status, Error> {
+    fn status(response: &Value) -> Result<u16, Error> {
         let status = response.funcall::<i64, _, _>("status", &[])?;
-        let status = u16::try_from(status).map_err(|_| Error::Status)?;
-        Status::from_code(status).ok_or(Error::Status)
+        u16::try_from(status).map_err(|_| Error::Status)
     }
 
     fn headers(response: &Value) -> Result<HashMap<String, String>, Error> {
@@ -81,14 +66,5 @@ impl Response {
             .flat_map(convert::identity)
             .collect::<Vec<_>>();
         Ok(bytes)
-    }
-}
-
-impl<'r> Responder<'r> for Error {
-    fn respond_to(self, _: &Request) -> response::Result<'r> {
-        response::Response::build()
-            .status(Status::InternalServerError)
-            .sized_body(Cursor::new(format!("{}", self)))
-            .ok()
     }
 }
